Walks and show your location Google maps
This projects allow user to enter the walks to the application. The
application will be able to store the array of walks into a file. This
application able to show open Firefox and launch the Google maps, and
using live maps showing your location as well as the location of the
walks. Using live maps instead of static image map, because using live
maps able to browse the nearby location relative to your location.